feat: add crouch filtering; prevent breaking bedrock

+++ src/main/java/me/radus/learningmod/event/MiningShapeEvents.java
@@ -1,22 +1,25 @@
 package me.radus.learningmod.event;
 
-import me.radus.learningmod.ModEnchantments;
 import me.radus.learningmod.util.MiningShapeHelpers;
 import net.minecraft.core.BlockPos;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.server.level.ServerPlayerGameMode;
-import net.minecraft.world.item.enchantment.EnchantmentHelper;
+import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraftforge.common.ForgeHooks;
 import net.minecraftforge.event.level.BlockEvent;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.common.Mod;
 
-import java.util.*;
+import java.util.HashSet;
+import java.util.Iterator;
+import java.util.Set;
+import java.util.UUID;
 
 @Mod.EventBusSubscriber(bus = Mod.EventBusSubscriber.Bus.FORGE)
 public class MiningShapeEvents {
-    private static final Set<UUID> currentlyMining = new HashSet<>();
+    private static final Set<UUID> playersCurrentlyMining = new HashSet<>();
 
     @SubscribeEvent
     public static void onBlockBreak(BlockEvent.BreakEvent event) {
@@ -25,7 +28,7 @@ public class MiningShapeEvents {
         }
 
         UUID playerId = player.getUUID();
-        if (currentlyMining.contains(playerId)) {
+        if (playersCurrentlyMining.contains(playerId)) {
             return;
         }
 
@@ -41,15 +44,27 @@ public class MiningShapeEvents {
             return;
         }
 
-        currentlyMining.add(playerId);
         BlockPos pos;
+        Level level = player.level();
+        playersCurrentlyMining.add(playerId);
 
         do {
             pos = breakableBlocks.next();
-            gameMode.destroyBlock(pos);
+            if (canBreakBlock(level, pos, player)) {
+                gameMode.destroyBlock(pos);
+            }
         } while (breakableBlocks.hasNext());
 
-        currentlyMining.remove(playerId);
+        playersCurrentlyMining.remove(playerId);
         event.setCanceled(true);
+    }
+
+    private static boolean canBreakBlock(Level level, BlockPos pos, Player player) {
+        if (!ForgeHooks.canEntityDestroy(level, pos, player)) {
+            return false;
+        }
+
+        BlockState state = level.getBlockState(pos);
+        return state.getDestroySpeed(level, pos) > 0.0f;
     }
 }
